pyproject.toml
[tool.poetry]
name = "fakes-n-contracts"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.8"
pydantic = { version = "^2.4.2", extras = ["email"] }
types-psycopg2 = "^2.9.21.14"
pytest-postgresql = "^5.0.0"
psycopg-binary = "^3.1.12"
aurora-data-api = "^0.4.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-icdiff = "^0.8"
pytest-clarity = "^1.0.1"
pytest-sugar = "^0.9.7"
polars = "^0.19.6"


[tool.poetry.group.tools.dependencies]
poethepoet = "^0.23.0"
ipython = "^8.16.0"
mypy = "^1.5.1"

[tool.poe.tasks]
main = { script = "src:main" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
